misc/aug/requirements.txt
ffmpeg==1.4
munch==4.0.0
numpy==2.1.3
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
